СС-39512: add @demo-smoke tags to AI workflows, audit logs, backoffice assistant, and cost price tests for improved categorization and filtering
