+++ cypress/fixtures/b2b-mp/demo/static-00-ai-configuration.json
